+++ apps/server/src/trpc/index.ts
@@ -6,11 +6,10 @@ import { settingsRouter } from './routes/settings';
 import { getContext } from 'hono/context-storage';
 import { draftsRouter } from './routes/drafts';
 import { labelsRouter } from './routes/label';
-import { brainRouter } from './routes/brain';
 import { notesRouter } from './routes/notes';
-import { voiceRouter } from './routes/voice';
-import { mailRouter } from './routes/mail';
+import { brainRouter } from './routes/brain';
 import { userRouter } from './routes/user';
+import { mailRouter } from './routes/mail';
 import type { HonoContext } from '../ctx';
 import { aiRouter } from './routes/ai';
 import { router } from './trpc';
@@ -27,7 +26,6 @@ export const appRouter = router({
   shortcut: shortcutRouter,
   settings: settingsRouter,
   user: userRouter,
-  voice: voiceRouter,
 });
 
 export type AppRouter = typeof appRouter;
