--- apps/server/src/trpc/routes/voice/index.ts
@@ -1,6 +1,0 @@
-import { getSignedUrl } from './get-signed-url';
-import { router } from '../../trpc';
-
-export const voiceRouter = router({
-  getSignedUrl,
-});
